requirements.txt
flask
flask_socketio
numpy==1.26.4
pandas-ta
python-binance
python-dotenv
setuptools